requirements.txt
fastapi[standard]==0.115.12
starlette==0.45.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
pydantic==2.11.2
pypika==0.48.9
pytest==8.3.4
testcontainers==4.10.0
uvicorn==0.34.0
h11~=0.14.0
pip~=21.1.2
wheel~=0.36.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.8
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
setuptools~=57.0.0
requests~=2.32.3
urllib3~=2.3.0
docker~=7.1.0
pluggy~=1.5.0
iniconfig~=2.1.0
certifi~=2025.1.31
anyio~=4.9.0
packaging~=24.2
PyYAML~=6.0.2
python-dotenv~=1.0.1
wrapt~=1.17.2
yandex-pgmigrate==1.0.9